% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmr_identification.R
\name{dmr_identification}
\alias{dmr_identification}
\title{DMR identification from DMCs}
\usage{
dmr_identification(x, diff_meth_cluster, data, legacy_data)
}
\arguments{
\item{x}{A \code{\link[betaHMM:betaHMM]{betaHMM}} object.}

\item{diff_meth_cluster}{The clusters identified as differentially methylated}

\item{data}{A dataframe of dimension \eqn{C \times NR}+2 containing the
chromosome, CpG site label and the methylation
values for \eqn{C} CpG sites from \eqn{R} treatment groups where
DNA samples are either collected from \eqn{N} patients
or has \eqn{N} replicates for each treatment group and this
dataset was passed as an argument to the
\code{\link[betaHMM:betaHMM]{betaHMM}} function.}

\item{legacy_data}{A dataset containing the manifest data
from the Illumina MethylationEPIC beadchip array.}
}
\value{
A dataframe containing the following columns:
\itemize{
\item start_CpG - The starting CpG site in the particular DMR
\item end_CpG -  The ending CpG site in the particular DMR
\item DMR_size - Number of CPG sites identified in the DMR
\item chr_dmr - The chromosome corresponding to the CpG sites in the DMR.
\item map_start - MAPINFO of starting CpG site in the particular DMR
\item map_end -  MAPINFO of the ending CpG site in the particular DMR}
}
\description{
Function to identify the DMRs for the DMCs identified using BHMM
}
\details{
Function to identify the DMRs for the DMCs identified using BHMM
}
\seealso{
\code{\link{betaHMM}}
}
