% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.betaHMM.R
\name{plot.betaHMM}
\alias{plot.betaHMM}
\title{Plots for visualizing the betaHMM class object}
\usage{
\method{plot}{betaHMM}(
  x,
  what = "fitted density",
  data = NULL,
  treatment_group = NULL,
  title = NULL,
  ...
)
}
\arguments{
\item{x}{A \code{\link[betaHMM:betaHMM]{betaHMM}} object.}

\item{what}{The different plots that can be obtained are either
"fitted density","kernel density" or
"uncertainty" (default = "fitted density").}

\item{data}{A dataframe of dimension \eqn{C \times NR} containing methylation
            values for \eqn{C} CpG sites from \eqn{R} treatment groups where
            DNA samples are either collected from \eqn{N} patients
            or has \eqn{N} replicates for each treatment group and this
            dataset was passed as an argument to the
            \code{\link[betaHMM:betaHMM]{betaHMM}} function.
The data is not required as an input when generating "uncertainty" plot and
 the default has been set as "NULL". The data needs to be passed as an
 argument to this function when generating either
 "fitted density" or "kernel density" plots.}

\item{treatment_group}{The names of the different treatment groups.
If no value is passed then default values of sample names, e.g. Sample 1,
Sample 2, etc are used as legend text (default = NULL).}

\item{title}{The title that the user wants to display.
If no title is to be displayed the default is "NULL".}

\item{...}{Other graphics parameters.}
}
\value{
This function displays the following plots as requested by the user:
\itemize{
\item fitted density estimates - Plot showing the fitted density estimates of the clustering solution under the optimal model selected.
\item kernel density estimates - Plot showing the kernel density estimates of the clustering solution under the optimal model selected.
\item uncertainty -  A boxplot showing the uncertainties in the optimal clustering solution.
}
}
\description{
Visualise a \code{\link[betaHMM:betaHMM]{betaHMM}} clustering solution by plotting the fitted and kernel density estimates and the uncertainty.
}
\details{
The fitted density estimates can be visualized under the optimal clustering solution by specifying what = "fitted density" and kernel density estimates under the optimal clustering solution by specifying what = "kernel density".
}
\seealso{
\code{\link{betaHMM}}
}
